+++ test/cmdlineTests/ir_compiler_inheritance_nosubobjects/input.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.6.0;
 
 contract C {