+++ test/cmdlineTests/require_overload/input.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.0;
 contract C {
     function f() public pure {