+++ test/libsolidity/syntaxTests/license/license_double.sol
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: GPL-3.0
+contract C {}
+// SPDX-License-Identifier: MIT
+// ----
+// ParserError: Multiple SPDX license identifiers found in source file. Use "AND" or "OR" to combine multiple licenses. Please see https://spdx.org for more information.